refactor: removed useless nesting inside /components where possible
