--- src/components/Button/index.tsx
@@ -1,1 +1,0 @@
-export { default as Button } from "@/components/Button/Button";
