--- src/components/Overlay/index.ts
@@ -1,1 +1,0 @@
-export { default as Overlay } from "@/components/Overlay/Overlay";
