--- src/components/Section/index.ts
@@ -1,1 +1,0 @@
-export { default as Section } from "@/components/Section/Section";
